refactor(Web Page): update imports of EditableContent in others components
